+++ daily_bets/db/nba_db.py
@@ -20,7 +20,7 @@ __all__: collections.abc.Sequence[str] = (
 
 import datetime
 import decimal
-import msgspec
+import msgspec  # noqa: F401
 import operator  # noqa: F401
 import typing
 
@@ -33,7 +33,7 @@ if typing.TYPE_CHECKING:
 
     ConnectionLike: typing.TypeAlias = asyncpg.Connection[asyncpg.Record] | asyncpg.pool.PoolConnectionProxy[asyncpg.Record]
 
-from daily_bets.db import models
+from daily_bets.db import models  # noqa: F401
 
 
 class NbaCopyAnalysisParams(msgspec.Struct):
@@ -74,7 +74,7 @@ WITH ranked AS (
             PARTITION BY
                 game_time,
                 game_tag,
-                (analysis->'input'->>'player_id')::int,
+                (analysis->'input'->>'player_id')::bigint,
                 analysis->'input'->>'stat',
                 (analysis->'input'->>'line')::numeric
             ORDER BY created_at DESC, id DESC
@@ -96,7 +96,7 @@ NBA_RECENT_ANALYSIS_KEYS: typing.Final[str] = """-- name: NbaRecentAnalysisKeys 
 SELECT
     game_time,
     game_tag,
-    (analysis->'input'->>'player_id')::int AS player_id,
+    (analysis->'input'->>'player_id')::bigint AS player_id,
     analysis->'input'->>'stat' AS stat,
     (analysis->'input'->>'line')::numeric AS line
 FROM public.v2_nba_daily_bets
@@ -117,8 +117,8 @@ WITH inserted AS (
         WHERE
             game_time = $3
             AND game_tag = $4
-            AND (analysis->'input'->>'player_id')::int =
-                ($1::json->'input'->>'player_id')::int
+            AND (analysis->'input'->>'player_id')::bigint =
+                ($1::json->'input'->>'player_id')::bigint
             AND analysis->'input'->>'stat' = ($1::json->'input'->>'stat')
             AND (analysis->'input'->>'line')::numeric =
                 ($1::json->'input'->>'line')::numeric
